Fix all places where gcc complains using strict compilation options. This is in preparation for C++11.

+++ src/XrdSfs/XrdSfsNative.hh
@@ -115,7 +115,7 @@ public:
         int            read(XrdSfsAio *aioparm);
 
         XrdSfsXferSize readv(XrdOucIOVec      *readV,
-                             size_t            readCount);
+                             int               readCount);
 
         XrdSfsXferSize write(XrdSfsFileOffset   fileOffset,
                              const char        *buffer,
